File Split (#80)
- Splitting all hpp files into cpp and hpp files. hpp files contain header information and were moved into the include folder, while cpp files contain the implementations and are built alongside main.cpp.
- Formatting changes--implementations now have the same or similar comments to what their header comments are.
