+++ CMakeLists.txt
@@ -5,12 +5,8 @@ project(project C CXX)
 add_subdirectory(include/json)
 add_subdirectory(include/httplib)
 
-
-add_executable(project main.cpp UMLParameter.cpp)
-
 # add executables here
-add_executable(project main.cpp)
-
+add_executable(project main.cpp UMLClass.cpp UMLAttribute.cpp UMLRelationship.cpp UMLData.cpp UMLFile.cpp UMLParameter.cpp CLI.cpp)
 
 # add threads for http library
 find_package (Threads REQUIRED)
@@ -20,8 +16,6 @@ target_link_libraries(project PUBLIC Threads::Threads)
 target_link_libraries(project PUBLIC json)
 
 target_link_libraries(project PUBLIC http)
-
-
 
 target_include_directories(project PUBLIC
                           "${PROJECT_BINARY_DIR}"
@@ -40,7 +34,8 @@ FetchContent_MakeAvailable(googletest)
 
 enable_testing()
 
-add_executable(Tests Tests.cpp UMLParameter.cpp)
+add_executable(Tests Tests.cpp UMLClass.cpp UMLAttribute.cpp UMLRelationship.cpp UMLData.cpp UMLFile.cpp UMLParameter.cpp CLI.cpp)
+
 target_link_libraries(Tests gtest_main)
 include(GoogleTest)
 gtest_discover_tests(Tests)
