+++ UMLFile.cpp
@@ -1,69 +1,20 @@
-#pragma once
 /*
-  Filename   : UMLFile.hpp
-  Description: Serves as an object for which information about a class
-  in the UML diagram is stored in a JSON file. 
+  Filename   : UMLFile.cpp
+  Description: Implementation of a file.
 */
-
 
 //--------------------------------------------------------------------
 // System includes
-#include <vector>
-#include <string>
-#include <fstream>
-#include "include/json/json.hpp"
-#include "UMLClass.hpp"
-#include "UMLData.hpp"
-#include "UMLAttribute.hpp"
-#include "UMLRelationship.hpp"
+#include "include/UMLFile.hpp"
 //--------------------------------------------------------------------
 
-//--------------------------------------------------------------------
-// Using declarations
-using std::string;
-using std::vector;
-using json = nlohmann::json;
-//--------------------------------------------------------------------
-
-class UMLFile
-{
-    private:
-
-        json jsonFile;
-        string path;
-
-    public:
-
-        //constructor: takes in the name of the file to save
-        UMLFile(const string&);
-
-        //saving file
-        void save(const UMLData& data);
-
-        //loads a system file and returns a UML data object
-        UMLData load();  
-
-    private:
-
-        //takes in vector of classes and adds a json string to the jsonFile
-        void addUMLClassVec(const vector<UMLClass>& umlclasses);
-
-        //takes in vector of relationships and adds them to the jsonFile
-        void addUMLRelationshipVec(const vector<UMLRelationship>& relationships);
-
-        //gets the classes from the json file and returns classes vector
-        vector<UMLClass> getUMLClassVec();
-        
-        //takes in UML Data object
-        //gets the relationships from the json file and adds them to the UMLData object
-        void getUMLRelationshipVec(UMLData& data);
-};
-
+// constructor: takes in the name of the file to save
 UMLFile::UMLFile(const string& newpath)
 :path(newpath)
 {
 }
 
+// saving file
 void UMLFile::save(const UMLData& data)
 {
     addUMLClassVec(data.getClasses());
@@ -75,6 +26,7 @@ void UMLFile::save(const UMLData& data)
     file.close();
 }
 
+// loads a system file and returns a UML data object
 UMLData UMLFile::load() 
 {
     std::ifstream file; 
@@ -88,6 +40,7 @@ UMLData UMLFile::load()
     return data;
 }
 
+// takes in vector of classes and adds a json string to the jsonFile
 void UMLFile::addUMLClassVec(const vector<UMLClass>& classes)
 {
     for (UMLClass umlclass : classes)
@@ -102,6 +55,7 @@ void UMLFile::addUMLClassVec(const vector<UMLClass>& classes)
     }    
 }
 
+// takes in vector of relationships and adds them to the jsonFile
 void UMLFile::addUMLRelationshipVec(const vector<UMLRelationship>& relationships)
 {
     for (UMLRelationship relationship : relationships)
@@ -110,6 +64,7 @@ void UMLFile::addUMLRelationshipVec(const vector<UMLRelationship>& relationships
     }    
 }
 
+// gets the classes from the json file and returns classes vector
 vector<UMLClass> UMLFile::getUMLClassVec()
 {
     vector<UMLClass> classVec;
@@ -126,6 +81,7 @@ vector<UMLClass> UMLFile::getUMLClassVec()
     return classVec;
 }
 
+// gets the relationships from the json file and adds them to the UMLData object
 void UMLFile::getUMLRelationshipVec(UMLData& data)
 {
     for (auto rel : jsonFile["Relationship"])
