+++ include/UMLAttribute.hpp
@@ -1,15 +1,10 @@
+#pragma once
 /*
   Filename   : UMLAttribute.hpp
   Description: Serves as an object for which information about an attribute
   in the UML diagram is stored. Stored as object for the sake of further
   implementation in later iterations.
 */
-
-/************************************************************/
-// Macro guard
-#ifndef UMLATTRIBUTE_HPP
-#define UMLATTRIBUTE_HPP
-/************************************************************/
 
 //--------------------------------------------------------------------
 // System includes
@@ -28,7 +23,6 @@ class UMLAttribute
 		string attributeName;
 
 	public:
-		// Params: string newAttribute (name of attribute)
 		// Constructor for attribute objects
 		UMLAttribute(string newAttribute);
 
@@ -39,22 +33,3 @@ class UMLAttribute
 		void changeName(string newAttributeName);
 
 };
-
-UMLAttribute::UMLAttribute(string newAttribute) 
-:attributeName(newAttribute)
-{
-}
-
-string UMLAttribute::getAttributeName() const
-{
-	return attributeName;
-}
-
-void UMLAttribute::changeName(string newAttributeName)
-{
-	attributeName = newAttributeName;
-}
-
-/************************************************************/
-#endif
-/************************************************************/
